Hints db fixes.
Hints DB now pulls genomes directly from the HAL to ensure correct order. Moved shared tasks to base_tasks.
