+++ CAT/base_tasks.py
@@ -219,3 +219,78 @@ class ToilTask(PipelineTask):
         """override the PipelineTask repr to report the batch system being used"""
         base_repr = super(ToilTask, self).__repr__()
         return 'Toil' + base_repr + ' using batchSystem {}'.format(self.batchSystem)
+
+
+class HintsDbTask(luigi.Task):
+    # path to config file
+    config = luigi.Parameter()
+    hal = luigi.Parameter()
+    augustus_hints_db = luigi.Parameter(default='augustus_hints.db')
+    work_dir = luigi.Parameter(default=os.path.join(tempfile.gettempdir(), __name__))
+    # Toil options
+    batchSystem = luigi.Parameter(default='singleMachine', significant=False)
+    maxCores = luigi.IntParameter(default=16, significant=False)
+    logLevel = luigi.Parameter(default='WARNING', significant=False)  # this is passed to toil
+    cleanWorkDir = luigi.Parameter(default='onSuccess', significant=False)  # debugging option
+    workDir = luigi.Parameter(default=None, significant=False)
+    disableCaching = luigi.BoolParameter(default=False, significant=False)
+    parasolCommand = luigi.Parameter(default=None, significant=False)
+    defaultMemory = luigi.IntParameter(default=8 * 1024 ** 3, significant=False)
+
+    def __repr__(self):
+        """override the repr to make logging cleaner"""
+        return 'HintsDbTask: {}'.format(self.__class__.__name__)
+
+
+class HintsDbWrapperTask(HintsDbTask, luigi.WrapperTask):
+    """add WrapperTask functionality to PipelineTask"""
+    pass
+
+
+class HintsDbToilTask(HintsDbTask):
+    """
+    Task for launching toil pipelines from within luigi.
+    """
+    resources = {'toil': 1}  # all toil pipelines use 1 toil
+
+    def prepare_toil_options(self, work_dir):
+        """
+        Prepares a Namespace object for Toil which has all defaults, overridden as specified
+        Will see if the jobStore path exists, and if it does, assume that we need to add --restart
+        :param work_dir: Parent directory where toil work will be done. jobStore will be placed inside. Will be used
+        to fill in the workDir class variable.
+        :return: Namespace
+        """
+        job_store = os.path.join(work_dir, 'jobStore')
+        tools.fileOps.ensure_file_dir(job_store)
+        toil_args = self.get_toil_defaults()
+        toil_args.__dict__.update(vars(self))
+        if os.path.exists(job_store):
+            try:
+                root_job = open(os.path.join(job_store, 'rootJobStoreID')).next().rstrip()
+                if not os.path.exists(os.path.join(job_store, 'tmp', root_job)):
+                    shutil.rmtree(job_store)
+                else:
+                    toil_args.restart = True
+            except OSError:
+                toil_args.restart = True
+            except IOError:
+                shutil.rmtree(job_store)
+        job_store = 'file:' + job_store
+        toil_args.jobStore = job_store
+        return toil_args
+
+    def get_toil_defaults(self):
+        """
+        Extracts the default toil options as a dictionary, setting jobStore to None
+        :return: dict
+        """
+        parser = Job.Runner.getDefaultArgumentParser()
+        namespace = parser.parse_args([''])  # empty jobStore attribute
+        namespace.jobStore = None  # jobStore attribute will be updated per-batch
+        return namespace
+
+    def __repr__(self):
+        """override the PipelineTask repr to report the batch system being used"""
+        base_repr = super(HintsDbToilTask, self).__repr__()
+        return 'Toil' + base_repr + ' using batchSystem {}'.format(self.batchSystem)
